# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobType = typeof JobType[keyof typeof JobType];


export const JobType = {
  transcription: 'transcription',
  subtitling: 'subtitling',
  captioning: 'captioning',
  dubbing: 'dubbing',
} as const;

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type JobDomain = typeof JobDomain[keyof typeof JobDomain];


export const JobDomain = {
  general: 'general',
  legal: 'legal',
  medical: 'medical',
  business: 'business',
} as const;

export interface Job {
  id: number;
  type: JobType;
  status: JobStatus;
  inputFilename: string;
  inputDurationMinutes: number;
  domain: JobDomain;
  /** @nullable */
  targetLanguage?: string | null;
  /** @nullable */
  outputFormat?: string | null;
  /** @nullable */
  outputUrl?: string | null;
  /** @nullable */
  creditsUsed?: number | null;
  /** @nullable */
  progressPercent?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type JobInputType = typeof JobInputType[keyof typeof JobInputType];


export const JobInputType = {
  transcription: 'transcription',
  subtitling: 'subtitling',
  captioning: 'captioning',
  dubbing: 'dubbing',
} as const;

export type JobInputDomain = typeof JobInputDomain[keyof typeof JobInputDomain];


export const JobInputDomain = {
  general: 'general',
  legal: 'legal',
  medical: 'medical',
  business: 'business',
} as const;

export interface JobInput {
  type: JobInputType;
  inputFilename: string;
  inputDurationMinutes: number;
  domain: JobInputDomain;
  targetLanguage?: string;
  outputFormat?: string;
}

export type JobStatusUpdateStatus = typeof JobStatusUpdateStatus[keyof typeof JobStatusUpdateStatus];


export const JobStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface JobStatusUpdate {
  status: JobStatusUpdateStatus;
  progressPercent?: number;
  outputUrl?: string;
  creditsUsed?: number;
  errorMessage?: string;
}

export type WalletPlanType = typeof WalletPlanType[keyof typeof WalletPlanType];


export const WalletPlanType = {
  free: 'free',
  pro: 'pro',
  firm: 'firm',
} as const;

export interface Wallet {
  balance: number;
  planType: WalletPlanType;
  totalSpent: number;
  totalJobsRun: number;
  totalMinutesProcessed?: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  credit: 'credit',
  debit: 'debit',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  /** @nullable */
  jobId?: number | null;
  createdAt: string;
}

export interface TopUpInput {
  amount: number;
}

export interface JobsByType {
  transcription: number;
  subtitling: number;
  captioning: number;
  dubbing: number;
}

export interface DashboardStats {
  totalJobs: number;
  completedJobs: number;
  failedJobs: number;
  processingJobs: number;
  totalMinutesProcessed: number;
  creditBalance: number;
  totalCreditsSpent: number;
  jobsByType: JobsByType;
}

export type TeamMemberRole = typeof TeamMemberRole[keyof typeof TeamMemberRole];


export const TeamMemberRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  role: TeamMemberRole;
  creditsUsed: number;
  joinedAt: string;
}

export type TeamMemberInviteRole = typeof TeamMemberInviteRole[keyof typeof TeamMemberInviteRole];


export const TeamMemberInviteRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface TeamMemberInvite {
  name: string;
  email: string;
  role: TeamMemberInviteRole;
}

export type ListJobsParams = {
type?: ListJobsType;
status?: ListJobsStatus;
limit?: number;
offset?: number;
};

export type ListJobsType = typeof ListJobsType[keyof typeof ListJobsType];


export const ListJobsType = {
  transcription: 'transcription',
  subtitling: 'subtitling',
  captioning: 'captioning',
  dubbing: 'dubbing',
} as const;

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type ListTransactionsParams = {
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};