# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */
import type { JobsByType } from './jobsByType';

export interface DashboardStats {
  totalJobs: number;
  completedJobs: number;
  failedJobs: number;
  processingJobs: number;
  totalMinutesProcessed: number;
  creditBalance: number;
  totalCreditsSpent: number;
  jobsByType: JobsByType;
}