# lib/api-zod/src/generated/types/jobStatusUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusUpdateStatus } from './jobStatusUpdateStatus';

export interface JobStatusUpdate {
  status: JobStatusUpdateStatus;
  progressPercent?: number;
  outputUrl?: string;
  creditsUsed?: number;
  errorMessage?: string;
}