# lib/api-zod/src/generated/types/jobsByType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */

export interface JobsByType {
  transcription: number;
  subtitling: number;
  captioning: number;
  dubbing: number;
}