# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMemberRole } from './teamMemberRole';

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  role: TeamMemberRole;
  creditsUsed: number;
  joinedAt: Date;
}