# lib/api-zod/src/generated/types/teamMemberInvite.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMemberInviteRole } from './teamMemberInviteRole';

export interface TeamMemberInvite {
  name: string;
  email: string;
  role: TeamMemberInviteRole;
}