# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DAK Transcription API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletPlanType } from './walletPlanType';

export interface Wallet {
  balance: number;
  planType: WalletPlanType;
  totalSpent: number;
  totalJobsRun: number;
  totalMinutesProcessed?: number;
}